requirements.txt
attrs==24.2.0
blinker==1.9.0
brotlicffi==1.1.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dnspython==2.7.0
flasgger==0.9.7.1
Flask==3.1.0
geographiclib==2.0
geopy==2.4.1
greenlet==2.0.2
gunicorn==23.0.0
httpbin==0.10.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
pycparser==2.22
pygeocoder==1.2.5
python-memcached==1.62
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.3
